Use single tactility.py and symbolic links
